requirements.txt
langchain==0.3.27
langchain-openai==0.2.17
langchain-community==0.3.27
langchain-core==0.3.74
langchain-chroma==0.2.5
langchain-huggingface==0.1.12
chromadb==1.0.17
openai==1.58.1
python-dotenv==0.9.9
fastapi==0.115.6
uvicorn==0.32.1
pydantic==2.10.4
tqdm==4.67.1

# Multi-modal and file handling
python-multipart==0.0.20
aiofiles==24.1.0
Pillow==10.4.0

# Document processing
PyPDF2==3.0.1
python-docx==1.1.2